feature(mockdata): Add support for custom configuration
