+++ .env-template
@@ -22,3 +22,6 @@ PORT="443"
 # IP addresses and "localhost" are *not* valid.
 # The whitelist must also include the domain defined above.
 QUICKFEED_WHITELIST="www.example.com,example.com"
+
+# Mockdata config
+QUICKFEED_MOCK_CONFIG=""
