+++ internal/mockdata/data.go
@@ -1,0 +1,145 @@
+package mockdata
+
+// TODO(Joachim): Sync all data with Github mock
+// TODO(Joachim): Create cache for generated data to read when setting up Github mock. (Dynamically generated data should be registered in the Github mock as well).
+
+const organization = "qf-2026"
+
+var (
+	defaultConfig = config{
+		Teachers:                        2,  // teachers + students = enrolledStudents
+		Students:                        10, // > enrolledStudents
+		EnrolledStudents:                8,  // < students
+		AssingnmentsPerCourse:           8,  // > groupAssignments
+		GroupAssignments:                2,  // < assingnmentsPerCourse
+		StudentSubmissionsPerAssignment: 8,
+		GroupSubmissionsPerAssignment:   3,
+		containerTimeout:                30,
+		Verbose:                         false,
+		Organizations:                   []string{organization},
+		Members:                         Members,
+		Courses: []string{
+			"Distributed Systems",
+			"Operating Systems",
+			"New Systems",
+			"Hyped Systems",
+		},
+		Groups: []string{
+			"Alpha Team",
+			"Beta Squad",
+			"Gamma Group",
+			"Delta Force",
+			"Epsilon Circle",
+			"Zeta Unit",
+			"Eta Crew",
+			"Theta Alliance",
+			"Iota Network",
+			"Kappa League",
+			"Lambda Syndicate",
+		},
+	}
+
+	// Members is a list of Members for each course.
+	Members = []string{
+		"Abigail Dyer",
+		"Max Clarkson",
+		"Irene Powell",
+		"Peter Jones",
+		"Jack Wallace",
+		"Dominic Grant",
+		"Amelia Henderson",
+		"Sonia Rutherford",
+		"Jack Marshall",
+		"Jennifer Howard",
+		"Carol Langdon",
+		"Zoe Chapman",
+		"Tracey Vaughan",
+		"Peter Alsop",
+		"Felicity Vaughan",
+		"Oliver Welch",
+		"Tim Harris",
+		"Vanessa Carr",
+		"Felicity Dowd",
+		"Frank Rees",
+		"Joan Ogden",
+		"Alexandra Manning",
+		"Kevin Underwood",
+		"Gavin Howard",
+		"Edward Dowd",
+		"Steven Peters",
+		"Liam Metcalfe",
+		"Trevor Howard",
+		"Dylan Miller",
+		"Stephanie Ball",
+		"Angela Poole",
+		"Samantha Slater",
+		"Thomas Smith",
+		"Justin Clarkson",
+		"Piers Cameron",
+		"Carolyn Hodges",
+		"William Wright",
+		"Christian Lawrence",
+		"Amy Kelly",
+		"Bella Dickens",
+		"Lauren Manning",
+		"Tracey Bailey",
+		"Gavin Stewart",
+		"Brandon Stewart",
+		"Diane Lyman",
+		"Joan Roberts",
+		"Anna Hill",
+		"Samantha Ross",
+		"Michael McLean",
+		"David Peters",
+		"Simon Reid",
+		"Andrew Greene",
+		"Diana MacLeod",
+		"Edward Murray",
+		"Molly MacDonald",
+		"Donna Walsh",
+		"Alan Taylor",
+		"Ruth Wallace",
+		"Madeleine Fisher",
+		"Una Springer",
+		"Cameron Slater",
+		"Peter Russell",
+		"Molly Black",
+		"Grace McLean",
+		"Bernadette Lawrence",
+		"Warren Johnston",
+		"Wendy Parsons",
+		"Steven Watson",
+		"Penelope Berry",
+		"Blake Powell",
+		"Theresa Wilson",
+		"Zoe Mackenzie",
+		"Samantha Tucker",
+		"Brian Dickens",
+		"Thomas Terry",
+		"Stephen Rutherford",
+		"Molly Brown",
+		"Anna Wright",
+		"Jacob Henderson",
+		"Irene Davies",
+		"Leah Simpson",
+		"Una Fisher",
+		"Molly Ogden",
+		"Victor Sutherland",
+		"Dominic Peters",
+		"Lucas Ross",
+		"Gavin Wilson",
+		"Stephen Bailey",
+		"Diana Lee",
+		"Phil Thomson",
+		"Alan Dickens",
+		"Vanessa James",
+		"Pippa Watson",
+		"Anne Lawrence",
+		"Anne Underwood",
+		"Anna White",
+		"Heather Clark",
+		"Karen Hudson",
+		"Una Powell",
+		"Owen Vance",
+	}
+)
