+++ internal/mockdata/groups.go
@@ -1,20 +1,21 @@
 package mockdata
 
 import (
-	"github.com/quickfeed/quickfeed/internal/qtest"
+	"fmt"
+
 	"github.com/quickfeed/quickfeed/qf"
 )
 
 func (g *generator) groups() error {
-	for i := 1; i <= courses; i++ {
-		id := uint64(g.Teachers() + 1)
+	for i := 1; i < len(g.Courses); i++ {
+		id := uint64(g.Teachers + 1)
 		groupStatus := qf.Group_PENDING
 		if i != 1 {
 			groupStatus = qf.Group_APPROVED
 		}
-		for j := range g.EnrolledStudents() / 2 {
+		for j := range g.EnrolledStudents / 2 {
 			group := &qf.Group{
-				Name:     qtest.Groups[j],
+				Name:     fmt.Sprintf("%s, %d", g.Groups[j], i),
 				CourseID: uint64(i),
 				Status:   groupStatus,
 				Users:    []*qf.User{{ID: id}, {ID: id + 1}},
